fix infinite loop bug with 256+ track points (happened with drift chambers)

+++ src/epecur/types.hpp
@@ -12,7 +12,7 @@ typedef	uint8_t	wire_id_ptr_t;
 BOOST_STATIC_ASSERT(sizeof(wire_id_ptr_t) >= sizeof(wire_id_t));
 
 typedef	double	wire_pos_t;
-typedef	uint8_t	wire_pos_ptr_t;
+typedef	uint16_t	wire_pos_ptr_t;
 
 typedef int8_t	small_angle_t;	// supposed to be from -90 to 90
